+++ src/pytransit/analysis/zinb.py
@@ -65,8 +65,8 @@ class ZinbMethod(base.MultiConditionMethod):
 
 
         combined_wig = args[0]
-        annotation = args[1]
-        metadata = args[2]
+        metadata = args[1]
+        annotation = args[2]
         output_file = args[3]
         normalization = kwargs.get("n", "TTR")
         NTerminus = float(kwargs.get("iN", 5.0))
@@ -416,7 +416,7 @@ class ZinbMethod(base.MultiConditionMethod):
 
     @classmethod
     def usage_string(self):
-        return """python %s zinb <combined wig file> <annotation .prot_table> <samples_metadata file> <output file> [Optional Arguments]
+        return """python %s zinb <combined wig file> <samples_metadata file> <annotation .prot_table> <output file> [Optional Arguments]
 
         Optional Arguments:
         -n <string>         :=  Normalization method. Default: -n TTR